術式防壁の大盾のテクスチャとモデル描画、及び一部共通演出として使っていたクールダウン明滅に関して共通化
