+++ src/main/resources/assets/apprenticecodex/geo/bulwark_greatshield.geo.json
@@ -74,14 +74,6 @@
 					]
 				},
 				{
-					"name": "core",
-					"parent": "parts",
-					"pivot": [0, 12.5, -0.5],
-					"cubes": [
-						{"origin": [-2, 10.5, -2], "size": [4, 4, 3], "inflate": -0.9, "uv": [8, 30]}
-					]
-				},
-				{
 					"name": "core_upper_base",
 					"parent": "parts",
 					"pivot": [0, 25, -0.5],
