+++ src/main/resources/assets/apprenticecodex/models/item/bulwark_greatshield.json
@@ -37,7 +37,7 @@
         -3
       ],
       "translation": [
-        5,
+        4,
         -27,
         0
       ],
@@ -54,7 +54,7 @@
         -3
       ],
       "translation": [
-        5,
+        4,
         -27,
         0
       ],
