+++ src/main/resources/assets/apprenticecodex/models/item/bulwark_greatshield_blocking.json
@@ -12,7 +12,7 @@
     },
     "firstperson_righthand": {
       "rotation": [0, 0, 3],
-      "translation": [-5, -19, 0],
+      "translation": [-5, -28, 0],
       "scale": [
         1.25,
         1.25,
@@ -21,7 +21,7 @@
     },
     "firstperson_lefthand": {
       "rotation": [0, 0, 3],
-      "translation": [-5, -19, 0],
+      "translation": [-5, -28, 0],
       "scale": [
         1.25,
         1.25,
